Split the frontend modules and fix every clippy lint
Completes the reorganization and the Rust best-practices pass. The workspace
now builds with zero warnings in debug, in release, and under
`cargo clippy --workspace --all-targets`, with the same tests passing
throughout (328 core + 8 real-ROM boot + 32 frontend, real ROMs included).

Frontend layout:
- `emulation/controller.rs` split into `snes.rs` (the `Cpu` + `SystemBus`
  composition and the region/frame-rate tables) and the controller proper,
  with its two test modules as siblings.
- `commands/library.rs` split into `library/{scan,store}` plus the command
  handlers. Note the constraint discovered here and written into CLAUDE.md:
  `invoke_handler!` resolves a companion macro from the command's own module
  path, so `#[tauri::command]` fns must stay in the module named there.

Encapsulation rather than public fields: the real-ROM tests used to reach
through `controller.snes.bus` and `snes.cpu.pc` directly. `Snes` now offers
`video_mode()` for the frame pacer plus `#[cfg(test)]` accessors
(`bus_mut`, `force_halt`, `program_counter`, `pc_and_accumulator`), so no
field had to be opened up to make the split compile.

Clippy, all fixed rather than allow-ed:
- `too_many_arguments` (6 sites). The renderer's draw calls took 10-11
  positional parameters; they now take a `Target` (the color buffer paired
  with its parallel layer-id buffer -- these must stay in lockstep, so they
  travel together), a `Frame` (the read-only VRAM/CGRAM/OAM/registers), and a
  `Band` (the y0..y1 range). `Voice::sample` took 13 and now takes a named
  `VoiceConfig`, which matters here specifically: an early version of this
  code had the register-to-field mapping offset by one slot, so `srcn` read
  VOL(L) and every voice played the wrong sample. Named fields make that
  class of mistake visible at the call site.
- `field_reassign_with_default` (25 sites) -> struct-update syntax, comments
  preserved per field.
- `needless_range_loop`, `unnecessary_cast`, `assign_op_pattern`,
  `identity_op`, `manual_rotate`, `manual_is_multiple_of`, `derivable_impls`,
  `ptr_arg` (`&PathBuf` -> `&Path`), `wrong_self_convention`
  (`Psw::to_byte` takes `self`, being `Copy`), `upper_case_acronyms`
  (`Country::USA` -> `Usa`, with `#[serde(rename)]` and its `as_str()` value
  unchanged so nothing user-visible moves), and the doc-comment lints -- two
  of which were `///` blocks describing a module while attached to a struct,
  now real `//!` module docs.
- `erasing_op`: `0 * SCREEN_WIDTH` written to mean "row 0".

Also removed a `mut` that only release builds warned about, by shadowing the
Tauri builder under `cfg(debug_assertions)` instead of reassigning it.

CLAUDE.md's architecture section described the old single-file layout, the
deleted `io.rs`/`keyboard.rs`, and the removed `windows` dependency; it now
describes the module tree, says which files are deliberately long and why,
and points at `set_video_mode` as the correct way to select NTSC/PAL.

Verified beyond the test suite: a real SMW frame rendered after the refactor
is statistically identical to one rendered before it (42 distinct colors,
mean luma 159.2) and was inspected by eye.

+++ OxideSFC/oxidesfc-core/src/bus/read.rs
@@ -47,7 +47,7 @@ impl SystemBus {
             // range, e.g. a plain `LDA $1234` with DB != 0 -- previously
             // unreachable only because nothing had run that far yet.
             if offset < 0x2000 {
-                let result = self.wram.read_u8(offset as u32)?;
+                let result = self.wram.read_u8(offset)?;
                 self.open_bus = result;
                 return Ok(result);
             }
